requirements.txt
fake-bpy-module==20251003
numpy==2.3.3
